+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/systems.js
@@ -230,7 +230,7 @@ if (!gwaioSystemChangesLoaded) {
                           " " +
                           Math.round(planet.generator.metalDensity)
                         );
-                    else if (planet.metal_spots)
+                    else if (planet.metal_spots && planet.generator.waterDepth)
                       return (
                         radius +
                         " " +
@@ -247,6 +247,42 @@ if (!gwaioSystemChangesLoaded) {
                         waterDepth +
                         " " +
                         Math.round(planet.generator.waterDepth) +
+                        "<br>" +
+                        waterHeight +
+                        " " +
+                        Math.round(planet.generator.waterHeight)
+                      );
+                    else if (planet.metal_spots)
+                      return (
+                        radius +
+                        " " +
+                        planet.generator.radius +
+                        "<br>" +
+                        metalSpots +
+                        " " +
+                        planet.metal_spots.length +
+                        "<br>" +
+                        temperature +
+                        " " +
+                        Math.round(planet.generator.temperature) +
+                        "<br>" +
+                        waterHeight +
+                        " " +
+                        Math.round(planet.generator.waterHeight)
+                      );
+                    else if (planet.generator.waterDepth)
+                      return (
+                        radius +
+                        " " +
+                        planet.generator.radius +
+                        "<br>" +
+                        metalSpots +
+                        " " +
+                        planet.metal_spots.length +
+                        "<br>" +
+                        temperature +
+                        " " +
+                        Math.round(planet.generator.temperature) +
                         "<br>" +
                         waterHeight +
                         " " +
@@ -269,10 +305,6 @@ if (!gwaioSystemChangesLoaded) {
                         temperature +
                         " " +
                         Math.round(planet.generator.temperature) +
-                        "<br>" +
-                        waterDepth +
-                        " " +
-                        Math.round(planet.generator.waterDepth) +
                         "<br>" +
                         waterHeight +
                         " " +